Move the confirm_report GUI presenter out of the validator service
protocol_validator.py imported confirm/YES from the pyface dialog wrapper,
which drags PySide6 into a service module at import time - the conventions
forbid Qt (direct or transitive) outside the view layer, and the module
docstring itself promised no Qt. The presenter and its _format_html /
_format_detail helpers now live in views/protocol_validator_presenter.py;
the service keeps validate_protocol + the headless log_report only.
